+++ packages/@tinacms/core/src/index.ts
@@ -17,6 +17,5 @@ limitations under the License.
 */
 
 export * from './cms'
-export * from './cms-forms/'
 export * from './subscribable'
 export * from './plugins'