+++ packages/@tinacms/fields/stories/wysiwyg.stories.tsx
@@ -35,4 +35,44 @@ const Basic = () => {
   )
 }
 
-storiesOf('Wysiwyg', module).add('Basic', () => <Basic />)
+const WithImage = () => {
+  const [value, setValue] = React.useState(
+    '![alt text](https://i.imgur.com/2FCfbgg.jpg "Logo Title Text 1")'
+  )
+  return (
+    <Wysiwyg
+      input={{
+        value,
+        onChange: (val: string) => {
+          console.log(val)
+          setValue(val)
+        },
+      }}
+    />
+  )
+}
+
+const WithTable = () => {
+  const [value, setValue] = React.useState(
+    `| A | B | C |
+| --- | --- | ---:|
+| aaa | bbb | ccc |
+| aaa | bbb | ccc |`
+  )
+  return (
+    <Wysiwyg
+      input={{
+        value,
+        onChange: (val: string) => {
+          console.log(val)
+          setValue(val)
+        },
+      }}
+    />
+  )
+}
+
+storiesOf('Wysiwyg', module)
+  .add('Basic', () => <Basic />)
+  .add('WithImage', () => <WithImage />)
+  .add('WithTable', () => <WithTable />)