+++ packages/@tinacms/form-builder/src/final-form-builder.tsx
@@ -17,7 +17,7 @@ limitations under the License.
 */
 
 import * as React from 'react'
-import { Form, Field } from '@tinacms/core'
+import { Form, Field } from '@tinacms/forms'
 import { useCMS } from '@tinacms/react-core'
 import {
   Form as FinalForm,
@@ -157,7 +157,7 @@ const FieldsGroup = styled.div`
   display: block;
   width: 100%;
   height: 100%;
-  padding: 1.25rem 1.25rem 0 1.25rem;
+  padding: 20px 20px 0 20px;
   white-space: nowrap;
   overflow-x: hidden;
   overflow-y: auto !important;