+++ packages/@tinacms/forms/types/react-frame-component.d.ts
@@ -16,19 +16,4 @@ limitations under the License.
 
 */
 
-import * as React from 'react'
-import { storiesOf } from '@storybook/react'
-import { ColorPicker } from '../src/ColorPicker'
-import { ColorFormat } from '../src/ColorPicker/color-formatter'
-
-storiesOf('ColorPicker', module).add('Hex', () => <BasicExample />)
-
-function BasicExample() {
-  const [value, setValue] = React.useState()
-  return (
-    <ColorPicker
-      colorFormat={ColorFormat.Hex}
-      input={{ value, onChange: setValue }}
-    />
-  )
-}
+declare module 'react-frame-component'