+++ packages/@tinacms/react-core/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.3...@tinacms/react-core@0.2.4-alpha.0) (2020-01-16)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
+## [0.2.3](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.2...@tinacms/react-core@0.2.3) (2020-01-14)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
+## [0.2.2](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.2-alpha.1...@tinacms/react-core@0.2.2) (2020-01-13)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
+## [0.2.2-alpha.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.2-alpha.0...@tinacms/react-core@0.2.2-alpha.1) (2020-01-10)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
 ## [0.2.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.1-alpha.1...@tinacms/react-core@0.2.1) (2019-12-17)
 
 **Note:** Version bump only for package @tinacms/react-core