+++ packages/gatsby-tinacms-remark/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.4-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.6.3...gatsby-tinacms-remark@0.6.4-alpha.0) (2020-01-16)
+
+
+### Bug Fixes
+
+* disregard intermittent null vals from gatsby ([d102745](https://github.com/tinacms/tinacms/commit/d102745))
+* make Field interface accept extra properties ([2a45c41](https://github.com/tinacms/tinacms/commit/2a45c41))
+
+
+
+
+
+## [0.6.3](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.6.2...gatsby-tinacms-remark@0.6.3) (2020-01-14)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
+## [0.6.2](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.6.2-alpha.1...gatsby-tinacms-remark@0.6.2) (2020-01-13)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
+## [0.6.2-alpha.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.6.2-alpha.0...gatsby-tinacms-remark@0.6.2-alpha.1) (2020-01-10)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
 ## [0.6.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.6.1-alpha.1...gatsby-tinacms-remark@0.6.1) (2019-12-17)
 
 **Note:** Version bump only for package gatsby-tinacms-remark